requirements.txt
GitPython==3.1.7
Sphinx==3.2.1
